bcmdhd4375: Downgrade the DPC thread from SCHED_FIFO to SCHED_NORMAL
When there is heavy WiFi traffic, the DPC thread can starve user processes
that are waiting for packets over WiFi because the DPC thread is a
real-time task, so it preempts all fair-scheduled tasks. This can result in
noticeable system-wide stutters.

Downgrade the DPC thread's scheduling class so that it is a fair-scheduled
task with a nice value of 0, which is the default for a freshly-created
kthread.

This makes the DPC thread behave more similarly to tasklets since ksoftirqd
is a fair-scheduled task with a nice value of 0.

Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>
Signed-off-by: Flopster101 <nahuelgomez329@gmail.com>